Update README.md
Digital Public Goods Webapp 
 
This repo hosts the code for the Digital Public Goods Alliance Webapp. 
At the user end, the Webapp provides a platform for open source digital products to submit their applications to be accepted as Digital Public Goods. 

At the back end, the Webapp provides a mechanism for the DPGA team to evaluate, accept, reject, and monitor the submissions. It allows reviewers to loop in experts for more nuanced opinions as well as get in touch with nominees for further clarifications. 

The Webapp keeps track of the number of DPG applications being processed per day and logs every action taken on a submitted application from a reviewer, expert or admin panel.   

Who have we made it for? - Users
We are building for 4 types of users
Applicants:
They are authorised representatives of digital projects who wish to get their project listed as a DPG.
DPGA Reviewers
These are members of the DPGA review team.
They review applications and make pass/fail decisions for each indicator.
Public Reviewers:
This could be anyone on the internet who is interested in contributing to the review of applications.
Admin:
These are members of the DPGA management team who manage the review process and ensure timely review of applications.
There will only be one admin account and access credentials will be shared among the relevant members.

How to contribute or submit feedback:

If you are a DPG applicant and have suggestions for the process, comment below so our team can look into the request and convert it into a pull request.